2016-03-31 Version 5.10.0 (Stable) Release
Notable changes:

* buffer:
  * make byteLength work with ArrayBuffer & DataView (Jackson Tian)
  * backport --zero-fill-buffers command line option (James M Snell)
  * backport new buffer constructor APIs (James M Snell)
  * add swap16() and swap32() methods (James M Snell)
* fs: add the fs.mkdtemp() function. (Florian MARGAINE)
* net: emit host in lookup event (HUANG Wei)
* node: --no-browser-globals configure flag (Fedor Indutny)
* npm: Upgrade to v3.8.3. Fixes a security flaw in the use of
authentication tokens in HTTP requests that would allow an attacker to
set up a server that could collect tokens from users of the command-line
interface. Authentication tokens have previously been sent with every
request made by the CLI for logged-in users, regardless of the destination
of the request. This update fixes this by only including those tokens
for requests made against the registry or registries used for the
current install. (Forrest L Norvell)
* repl: support standalone blocks (Prince J Wesley)
* src: override v8 thread defaults using cli options (Tom Gallacher)

Signed-off-by: Ilkka Myller <ilkka.myller@nodefield.com>